AssessmentTask/requirements.txt
googlemaps==3.1.4
django_rest_swagger==2.2.0
Django==3.0
djangorestframework==3.11.0
django-rest-auth
django-allauth
pymysql

